# perf: store/cachekv: avoid a map lookup if unnecessary, clear maps fast, avoid keys sort