@odeke-em 2021-11-08
> ACK, can you add a changelog entry?

Done, thank you @fedekunze!